Crate amethyst_rendy

Re-exports

pub use palette;

pub use rendy;

pub use formats::mesh::MeshPrefab;

pub use formats::texture::TexturePrefab;

pub use mtl::Material;

pub use mtl::MaterialDefaults;

pub use sprite::Sprite;

pub use sprite::SpriteRender;

pub use sprite::SpriteSheet;

pub use system::GraphCreator;

pub use system::RenderingSystem;

pub use types::Backend;

pub use types::Mesh;

pub use types::Texture;

pub use util::simple_shader_set;

pub use util::ChangeDetection;

Modules

batch

camera
Camera type with support for perspective and orthographic projections.

debug_drawing

error
Renderer error types.

formats

light
Light sources.

loaders

mtl
Physically-based material.

pass

pipeline

pod

resources
amethyst rendering ecs resources

serde_shim
This module contains a few serialization shims for types from palette crate. It is required due to buggy default implementation of those types. See this issue for more details: https://github.com/Ogeon/palette/issues/130 When above issue will be resolved, this can probably be removed.

shape

skinning

sprite

sprite_visibility

submodules

system
Renderer system

transparent

types

util

visibility