Module amethyst_rendy::batch

Module containing structures useful for batching draw calls in scenarios with various known assumptions, e.g. order independence.

Structs

OneLevelBatch
A batching implementation with one level of indexing. Data type D batched by primary key PK. Items with the same PK are always grouped.

OrderedOneLevelBatch
A batching implementation with one level of indexing. Data type D batched by primary key PK.

OrderedTwoLevelBatch
Batching implementation which provides two levels of indirection and grouping for a given batch. This batch method is used, for example, batching meshes and textures; for any given draw call, a user would want to batch all draws using a specific texture together, and then also group all draw calls for a specific mesh together.

TwoLevelBatch
Batching implementation which provides two levels of indirection and grouping for a given batch. This batch method is used, for example, batching meshes and textures; for any given draw call, a user would want to batch all draws using a specific texture together, and then also group all draw calls for a specific mesh together.

Traits

GroupIterator
Iterator trait for grouping iterated 2-tuples (K, V) by contiguous ranges with equal K, providing access in a group-by-group manner.