Module amethyst_rendy::camera

Camera type with support for perspective and orthographic projections.

Structs

ActiveCamera
Active camera resource, used by the renderer to choose which camera to get the view matrix from. If no active camera is found, the first camera will be used as a fallback.

ActiveCameraPrefab
Active camera prefab

Camera
Camera struct.

Orthographic

Perspective

Enums

CameraPrefab
Projection prefab

Projection
The projection mode of a Camera.