Module amethyst_rendy::debug_drawing

Debug Drawing library

Structs

DebugLine
Debug lines are stored as a pair of position and color.

DebugLines
Resource that stores non-persistent debug lines to be rendered in DebugLinesPass draw pass. The vector is automatically cleared after being rendered.

DebugLinesComponent
Component that stores persistent debug lines to be rendered in DebugLinesPass draw pass. The vector can only be cleared manually.

DebugLinesParams
Parameters for renderer of debug lines. The params affect all lines.