Module amethyst_rendy::mtl

Physically-based material.

Structs

Material
A physically based Material with metallic workflow, fully utilized in PBR render pass.

MaterialDefaults
A resource providing default textures for Material. These will be be used by the renderer in case a texture handle points to a texture which is not loaded already. Additionally, you can use it to fill up the fields of Material you don't want to specify.

TexAlbedo
Macro Generated Texture Type

TexAmbientOcclusion
Macro Generated Texture Type

TexCavity
Macro Generated Texture Type

TexEmission
Macro Generated Texture Type

TexMetallicRoughness
Macro Generated Texture Type

TexNormal
Macro Generated Texture Type

TextureOffset
Material reference this part of the texture

Traits

StaticTextureSet
Trait providing generic access to a collection of texture handles

Type Definitions

FullTextureSet
Type alias for a tuple collection of a complete PBR texture set.