Enum amethyst_rendy::palette::Color

pub enum Color<S = Srgb, T = f32> where
    S: RgbSpace,
    T: Float + Component,  {
    Luma(Luma<Linear<<S as RgbSpace>::WhitePoint>, T>),
    Rgb(Rgb<Linear<S>, T>),
    Xyz(Xyz<<S as RgbSpace>::WhitePoint, T>),
    Yxy(Yxy<<S as RgbSpace>::WhitePoint, T>),
    Lab(Lab<<S as RgbSpace>::WhitePoint, T>),
    Lch(Lch<<S as RgbSpace>::WhitePoint, T>),
    Hsv(Hsv<S, T>),
    Hsl(Hsl<S, T>),
    Hwb(Hwb<S, T>),
}

A generic color type.

The Color may belong to any color space and it may change depending on which operation is performed. That makes it immune to the "without conversion" rule of the operations it supports. The color spaces are selected as follows:

Mix: RGB for no particular reason, except that it's intuitive.

Shade: CIE L*a*b* for its luminance component.

Hue and GetHue: CIE L*C*h° for its hue component and how it preserves the apparent lightness.

Saturate: CIE L*C*h° for its chromaticity component and how it preserves the apparent lightness.

It's not recommended to use Color when full control is necessary, but it can easily be converted to a fixed color space in those cases.

Variants

Luma(Luma<Linear<<S as RgbSpace>::WhitePoint>, T>)

Linear luminance.

Rgb(Rgb<Linear<S>, T>)

Linear RGB.

Xyz(Xyz<<S as RgbSpace>::WhitePoint, T>)

CIE 1931 XYZ.

Yxy(Yxy<<S as RgbSpace>::WhitePoint, T>)

CIE 1931 Yxy.

Lab(Lab<<S as RgbSpace>::WhitePoint, T>)

CIE L*a*b* (CIELAB).

Lch(Lch<<S as RgbSpace>::WhitePoint, T>)

CIE L*C*h°, a polar version of CIE L*a*b*.

Hsv(Hsv<S, T>)

Linear HSV, a cylindrical version of RGB.

Hsl(Hsl<S, T>)

Linear HSL, a cylindrical version of RGB.

Hwb(Hwb<S, T>)

Linear HWB, an intuitive cylindrical version of RGB.

Implementations

impl<T> Color<Srgb, T> where
T: Component + Float,

pub fn linear_y(luma: T) -> Color<Srgb, T>

Linear luminance.

pub fn linear_rgb(red: T, green: T, blue: T) -> Color<Srgb, T>

Linear RGB.

pub fn xyz(x: T, y: T, z: T) -> Color<Srgb, T>

CIE XYZ.

pub fn yxy(x: T, y: T, luma: T) -> Color<Srgb, T>

CIE Yxy.

pub fn lab(l: T, a: T, b: T) -> Color<Srgb, T>

CIE L*a*b*.

pub fn lch(l: T, chroma: T, hue: LabHue<T>) -> Color<Srgb, T>

CIE L*C*h°.

pub fn hsv(hue: RgbHue<T>, saturation: T, value: T) -> Color<Srgb, T>

Linear HSV.

pub fn hsl(hue: RgbHue<T>, saturation: T, lightness: T) -> Color<Srgb, T>

Linear HSL.

pub fn hwb(hue: RgbHue<T>, whiteness: T, balckness: T) -> Color<Srgb, T>

Linear HWB.