Struct amethyst_rendy::palette::Lab

#[repr(C)]pub struct Lab<Wp = D65, T = f32> where
    T: Component + Float,
    Wp: WhitePoint,  {
    pub l: T,
    pub a: T,
    pub b: T,
    pub white_point: PhantomData<Wp>,
}

The CIE L*a*b* (CIELAB) color space.

CIE L*a*b* is a device independent color space which includes all perceivable colors. It's sometimes used to convert between other color spaces, because of its ability to represent all of their colors, and sometimes in color manipulation, because of its perceptual uniformity. This means that the perceptual difference between two colors is equal to their numerical difference.

The parameters of L*a*b* are quite different, compared to many other color spaces, so manipulating them manually may be unintuitive.

Fields
l: T

L* is the lightness of the color. 0.0 gives absolute black and 100 give the brightest white.

a: T

a* goes from red at -128 to green at 127.

b: T

b* goes from yellow at -128 to blue at 127.

white_point: PhantomData<Wp>

The white point associated with the color's illuminant and observer. D65 for 2 degree observer is used by default.

Implementations

impl<T> Lab<D65, T> where
T: Component + Float,

pub fn new(l: T, a: T, b: T) -> Lab<D65, T>

CIE L*a*b* with white point D65.

impl<Wp, T> Lab<Wp, T> where
T: Component + Float,
Wp: WhitePoint,

pub fn with_wp(l: T, a: T, b: T) -> Lab<Wp, T>

CIE L*a*b*.

pub fn into_components(self) -> (T, T, T)

Convert to a (L\*, a\*, b\*) tuple.

pub fn from_components((T, T, T)) -> Lab<Wp, T>

Convert from a (L\*, a\*, b\*) tuple.