Module amethyst_rendy::rendy

Rendy's top level crate. Reexports all others.

Modules

command
This crate revolves around command recording and submission.

descriptor

factory
Higher-level rendy interface.

frame
Everything that is frame related.

graph
Framegraph implementation for Rendy engine.

hal
Low-level graphics abstraction for Rust. Mostly operates on data, not types. Designed for use by libraries and higher-level abstractions only.

memory
GPU memory management

mesh
This crates provides means to deal with vertex buffers and meshes.

resource
This crate provide methods to create/destroy and otherwise manage device resources. Primarily focus on buffers and images.

shader
Shader compilation.

texture
Texture creation and usage

util
Crate that contains utility modules used by other rendy crates

wsi
Window system integration.