Function amethyst_rendy::rendy::command::families_from_device

pub unsafe fn families_from_device<B, impl IntoIterator, impl gfx_hal::queue::QueueFamily>(
    device: DeviceId, 
    queues: &mut Queues<B>, 
    families: impl IntoIterator, 
    queue_types: &[impl gfx_hal::queue::QueueFamily]
) -> Families<B> where
    B: Backend,
    impl IntoIterator: IntoIterator<Item = (FamilyId, usize)>,
    impl gfx_hal::queue::QueueFamily: QueueFamily, 

Query queue families from device.

Safety

This function shouldn't be used more then once with same parameters. Raw queue handle queried from device can make returned Family usage invalid. families iterator must yeild unique family indices with queue count used during device creation. properties must contain properties retuned for queue family from physical device for each family id yielded by families.