Module amethyst_rendy::rendy::hal::adapter

Physical devices and adapters.

The PhysicalDevice trait specifies the API a backend must provide for dealing with and querying a physical device, such as a particular GPU. An Adapter is a struct containing a PhysicalDevice and metadata for a particular GPU, generally created from an Instance of that backend. adapter.open_with(...) will return a Device that has the properties specified.

Structs

Adapter
The list of Adapter instances is obtained by calling Instance::enumerate_adapters().

AdapterInfo
Metadata about a backend adapter.

MemoryProperties
Types of memory supported by this adapter and available memory.

MemoryType
A description for a single chunk of memory in a heap.

MemoryTypeId
A strongly-typed index to a particular MemoryType.

Enums

DeviceType
Supported physical device types

Traits

PhysicalDevice
Represents a physical device (such as a GPU) capable of supporting the given backend.

Type Definitions

QueuePriority
Scheduling hint for devices about the priority of a queue. Values range from 0.0 (low) to 1.0 (high).