Struct amethyst_rendy::rendy::hal::command::RenderPassInlineEncoder

pub struct RenderPassInlineEncoder<'a, B>(_)
 where
    B: Backend,
    <B as Backend>::CommandBuffer: 'a;

An object that records commands into a command buffer inline, that is, without secondary command buffers.

Implementations

impl<'a, B> RenderPassInlineEncoder<'a, B> where
B: Backend,

pub unsafe fn new<C, T, S>(
cmd_buffer: &'a mut CommandBuffer<B, C, S, Primary, <B as Backend>::CommandBuffer>,
render_pass: &<B as Backend>::RenderPass,
frame_buffer: &<B as Backend>::Framebuffer,
render_area: Rect,
clear_values: T
) -> RenderPassInlineEncoder<'a, B> where
C: Supports<Graphics>,
S: Shot,
T: IntoIterator,
<T as IntoIterator>::Item: Borrow<ClearValue>,

Creates a new RenderPassInlineEncoder, starting a new render pass in the given CommandBuffer.

pub fn next_subpass_inline(self) -> RenderPassInlineEncoder<'a, B>

Start the next subpass.

pub fn next_subpass_secondary(self) -> RenderPassSecondaryEncoder<'a, B>

Begins recording a new subpass with secondary buffers.