Module amethyst_rendy::rendy::hal::queue

Command queues.

Queues are the execution paths of the graphical processing units. These process submitted commands buffers.

There are different types of queues, which can only handle associated command buffers. CommandQueue<B, C> has the capability defined by C: graphics, compute and transfer.

Modules

capability
Type system encoded queue capabilities.

family
Queue family and groups.

Structs

CommandQueue
Stronger-typed and safer CommandQueue wraps around RawCommandQueue.

QueueFamilyId
Identifier for a queue family of a physical device.

QueueGroup
Strong-typed group of queues of the same queue family.

Queues
Contains a list of all instantiated queues. Conceptually structured as a collection of QueueGroups, one for each queue family.

Submission
Submission information for a command queue.

Enums

Compute
Compute capability, supporting compute and transfer operations.

General
General capability, supporting graphics, compute and transfer operations.

Graphics
Graphics capability, supporting graphics and transfer operations.

QueueType
The type of the queue, an enum encompassing queue::Capability

Transfer
Transfer capability, supporting only transfer operations.

Traits

Capability
A Capability is an object that specifies what kind of operations a queue type performs, allowing what types support what queue operations to be described at runtime by the type system.

QueueFamily
General information about a queue family, available upon adapter discovery.

RawCommandQueue
RawCommandQueue are abstractions to the internal GPU execution engines. Commands are executed on the the device by submitting command buffers to queues.

Supports
A trait that indicates that a particular type of queue supports a particular Capability.