Trait amethyst_rendy::rendy::hal::queue::family::QueueFamily

pub trait QueueFamily: Send + Sync + Any + Debug {
    fn queue_type(&self) -> QueueType;

    fn max_queues(&self) -> usize;

    fn id(&self) -> QueueFamilyId;

    fn supports_graphics(&self) -> bool { ... }

    fn supports_compute(&self) -> bool { ... }

    fn supports_transfer(&self) -> bool { ... }
}

General information about a queue family, available upon adapter discovery.

Note that a backend can expose multiple queue families with the same properties.

Required methods

fn queue_type(&self) -> QueueType

Returns the type of queues.

fn max_queues(&self) -> usize

Returns maximum number of queues created from this family.

fn id(&self) -> QueueFamilyId

Returns the queue family ID.

Loading content...
Provided methods

fn supports_graphics(&self) -> bool

Returns true if the queue supports graphics operations.

fn supports_compute(&self) -> bool

Returns true if the queue supports compute operations.

fn supports_transfer(&self) -> bool

Returns true if the queue supports transfer operations.

Loading content...