Struct amethyst_rendy::rendy::hal::Limits

pub struct Limits {
    pub max_image_1d_size: u32,
    pub max_image_2d_size: u32,
    pub max_image_3d_size: u32,
    pub max_image_cube_size: u32,
    pub max_image_array_layers: u16,
    pub max_texel_elements: usize,
    pub max_uniform_buffer_range: u64,
    pub max_storage_buffer_range: u64,
    pub max_push_constants_size: usize,
    pub max_memory_allocation_count: usize,
    pub max_sampler_allocation_count: usize,
    pub max_bound_descriptor_sets: u16,
    pub max_framebuffer_layers: usize,
    pub max_per_stage_descriptor_samplers: usize,
    pub max_per_stage_descriptor_uniform_buffers: usize,
    pub max_per_stage_descriptor_storage_buffers: usize,
    pub max_per_stage_descriptor_sampled_images: usize,
    pub max_per_stage_descriptor_storage_images: usize,
    pub max_per_stage_descriptor_input_attachments: usize,
    pub max_per_stage_resources: usize,
    pub max_descriptor_set_samplers: usize,
    pub max_descriptor_set_uniform_buffers: usize,
    pub max_descriptor_set_uniform_buffers_dynamic: usize,
    pub max_descriptor_set_storage_buffers: usize,
    pub max_descriptor_set_storage_buffers_dynamic: usize,
    pub max_descriptor_set_sampled_images: usize,
    pub max_descriptor_set_storage_images: usize,
    pub max_descriptor_set_input_attachments: usize,
    pub max_vertex_input_attributes: usize,
    pub max_vertex_input_bindings: usize,
    pub max_vertex_input_attribute_offset: usize,
    pub max_vertex_input_binding_stride: usize,
    pub max_vertex_output_components: usize,
    pub max_patch_size: u8,
    pub max_geometry_shader_invocations: usize,
    pub max_geometry_input_components: usize,
    pub max_geometry_output_components: usize,
    pub max_geometry_output_vertices: usize,
    pub max_geometry_total_output_components: usize,
    pub max_fragment_input_components: usize,
    pub max_fragment_output_attachments: usize,
    pub max_fragment_dual_source_attachments: usize,
    pub max_fragment_combined_output_resources: usize,
    pub max_compute_shared_memory_size: usize,
    pub max_compute_work_group_count: [u32; 3],
    pub max_compute_work_group_invocations: usize,
    pub max_compute_work_group_size: [u32; 3],
    pub max_draw_indexed_index_value: u32,
    pub max_draw_indirect_count: u32,
    pub max_sampler_lod_bias: f32,
    pub max_sampler_anisotropy: f32,
    pub max_viewports: usize,
    pub max_viewport_dimensions: [u32; 2],
    pub max_framebuffer_extent: Extent,
    pub min_memory_map_alignment: usize,
    pub buffer_image_granularity: u64,
    pub min_texel_buffer_offset_alignment: u64,
    pub min_uniform_buffer_offset_alignment: u64,
    pub min_storage_buffer_offset_alignment: u64,
    pub framebuffer_color_sample_counts: u8,
    pub framebuffer_depth_sample_counts: u8,
    pub framebuffer_stencil_sample_counts: u8,
    pub max_color_attachments: usize,
    pub standard_sample_locations: bool,
    pub optimal_buffer_copy_offset_alignment: u64,
    pub optimal_buffer_copy_pitch_alignment: u64,
    pub non_coherent_atom_size: usize,
    pub min_vertex_input_binding_stride_alignment: u64,
}

Resource limits of a particular graphics device.

Fields
max_image_1d_size: u32

Maximum supported image 1D size.

max_image_2d_size: u32

Maximum supported image 2D size.

max_image_3d_size: u32

Maximum supported image 3D size.

max_image_cube_size: u32

Maximum supported image cube size.

max_image_array_layers: u16

Maximum supporter image array size.

max_texel_elements: usize

Maximum number of elements for the BufferView to see.

max_uniform_buffer_range: u64

max_storage_buffer_range: u64

max_push_constants_size: usize

max_memory_allocation_count: usize

max_sampler_allocation_count: usize

max_bound_descriptor_sets: u16

max_framebuffer_layers: usize

max_per_stage_descriptor_samplers: usize

max_per_stage_descriptor_uniform_buffers: usize

max_per_stage_descriptor_storage_buffers: usize

max_per_stage_descriptor_sampled_images: usize

max_per_stage_descriptor_storage_images: usize

max_per_stage_descriptor_input_attachments: usize

max_per_stage_resources: usize

max_descriptor_set_samplers: usize

max_descriptor_set_uniform_buffers: usize

max_descriptor_set_uniform_buffers_dynamic: usize

max_descriptor_set_storage_buffers: usize

max_descriptor_set_storage_buffers_dynamic: usize

max_descriptor_set_sampled_images: usize

max_descriptor_set_storage_images: usize

max_descriptor_set_input_attachments: usize

max_vertex_input_attributes: usize

Maximum number of vertex input attributes that can be specified for a graphics pipeline.

max_vertex_input_bindings: usize

Maximum number of vertex buffers that can be specified for providing vertex attributes to a graphics pipeline.

max_vertex_input_attribute_offset: usize

Maximum vertex input attribute offset that can be added to the vertex input binding stride.

max_vertex_input_binding_stride: usize

Maximum vertex input binding stride that can be specified in a vertex input binding.

max_vertex_output_components: usize

Maximum number of components of output variables which can be output by a vertex shader.

max_patch_size: u8

Maximum number of vertices for each patch.

max_geometry_shader_invocations: usize

max_geometry_input_components: usize

max_geometry_output_components: usize

max_geometry_output_vertices: usize

max_geometry_total_output_components: usize

max_fragment_input_components: usize

max_fragment_output_attachments: usize

max_fragment_dual_source_attachments: usize

max_fragment_combined_output_resources: usize

max_compute_shared_memory_size: usize

max_compute_work_group_count: [u32; 3]

max_compute_work_group_invocations: usize

max_compute_work_group_size: [u32; 3]

max_draw_indexed_index_value: u32

max_draw_indirect_count: u32

max_sampler_lod_bias: f32

max_sampler_anisotropy: f32

Maximum degree of sampler anisotropy.

max_viewports: usize

Maximum number of viewports.

max_viewport_dimensions: [u32; 2]

max_framebuffer_extent: Extent

min_memory_map_alignment: usize

buffer_image_granularity: u64

min_texel_buffer_offset_alignment: u64

The alignment of the start of buffer used as a texel buffer, in bytes, non-zero.

min_uniform_buffer_offset_alignment: u64

The alignment of the start of buffer used for uniform buffer updates, in bytes, non-zero.

min_storage_buffer_offset_alignment: u64

The alignment of the start of buffer used as a storage buffer, in bytes, non-zero.

framebuffer_color_sample_counts: u8

Number of samples supported for color attachments of framebuffers (floating/fixed point).

framebuffer_depth_sample_counts: u8

Number of samples supported for depth attachments of framebuffers.

framebuffer_stencil_sample_counts: u8

Number of samples supported for stencil attachments of framebuffers.

max_color_attachments: usize

Maximum number of color attachments that can be used by a subpass in a render pass.

standard_sample_locations: bool

optimal_buffer_copy_offset_alignment: u64

The alignment of the start of the buffer used as a GPU copy source, in bytes, non-zero.

optimal_buffer_copy_pitch_alignment: u64

The alignment of the row pitch of the texture data stored in a buffer that is used in a GPU copy operation, in bytes, non-zero.

non_coherent_atom_size: usize

Size and alignment in bytes that bounds concurrent access to host-mapped device memory.

min_vertex_input_binding_stride_alignment: u64

The alignment of the vertex buffer stride.