Trait amethyst_rendy::rendy::hal::Capability

pub trait Capability {
    fn supported_by(qt: QueueType) -> bool;
}

A Capability is an object that specifies what kind of operations a queue type performs, allowing what types support what queue operations to be described at runtime by the type system.

Required methods

fn supported_by(qt: QueueType) -> bool

Return true if this type level capability is supported by a run-time queue type.

Loading content...