Module amethyst_rendy::rendy::mesh

This crates provides means to deal with vertex buffers and meshes.

Attribute and VertexFormat allow vertex structure to declare semantics. Mesh can be created from typed vertex structures and provides mechanism to bind vertex attributes required by shader interface.

Modules

obj
Loading mesh data from obj format.

Structs

Incompatible
failure::Error type returned by Mesh::bind in case of mesh's vertex buffers are incompatible with requested vertex formats.

IndexBuffer
Index buffer with it's type

Mesh
Single mesh is a collection of buffer ranges that provides available attributes. Usually exactly one mesh is used per draw call.

MeshBuilder
Generics-free mesh builder.

VertexBufferLayout
Vertex buffer with it's format

Enums

Indices
Abstracts over two types of indices and their absence.