Module amethyst_rendy::rendy::resource

This crate provide methods to create/destroy and otherwise manage device resources. Primarily focus on buffers and images.

Structs

Buffer
Generic buffer resource wrapper.

BufferInfo
Buffer info.

DescriptorSet
Generic descriptor set resource wrapper.

DescriptorSetInfo
Descriptor set layout info.

DescriptorSetLayout
Generic descriptor set layout resource wrapper.

Escape
Allows values to "escape" dropping by sending them to the Terminal.

Handle
Allows values to "escape" dropping by sending them to the Terminal. Permit sharing unlike Escape

Image
Generic image resource wrapper.

ImageInfo
Image info.

ImageView
Generic image view resource wrapper.

ImageViewInfo
Image view info

ResourceTracker
Resource handler.

Sampler
Generic sampler resource wrapper.

SamplerCache
Sampler cache holds handlers to created samplers.

Terminal
This types allows the user to create Escape wrappers. Receives values from dropped Escape instances that was created by this Terminal.