Module amethyst_rendy::rendy::texture::palette

Module that generates functions to load a single color into a 1x1 Texture. Which function to use depends on what color space the user is using.

Functions

load_from_linear_luma
Function to load texture from palette pixels.

load_from_linear_luma_f32
Function to load texture from palette pixels.

load_from_linear_lumaa
Function to load texture from palette pixels.

load_from_linear_lumaa_f32
Function to load texture from palette pixels.

load_from_linear_rgb
Function to load texture from palette pixels.

load_from_linear_rgb_f32
Function to load texture from palette pixels.

load_from_linear_rgb_u16
Function to load texture from palette pixels.

load_from_linear_rgba
Function to load texture from palette pixels.

load_from_linear_rgba_f32
Function to load texture from palette pixels.

load_from_linear_rgba_u16
Function to load texture from palette pixels.

load_from_luma
Function to load texture from palette pixels.

load_from_lumaa
Function to load texture from palette pixels.

load_from_srgb
Function to load texture from palette pixels.

load_from_srgba
Function to load texture from palette pixels.