Module amethyst_rendy::serde_shim::srgb

Srgb serialization shim.

#[derive(Serialize, Deserialize)]
struct MyType(
    #[serde(with="amethyst_rendy::serde_shim::srgb")]
    pub palette::Srgb
);

Functions

deserialize
Deserialize Srgba type as tuple struct with three floats

serialize
Serialize Srgba type as tuple struct with three floats