Trait amethyst_rendy::shape::FromInternalVertex

pub trait FromInternalVertex {
    fn from_internal(v: &InternalVertexData) -> Self;
}

Internal trait for converting from vertex data to a shape type.

Required methods

fn from_internal(v: &InternalVertexData) -> Self

Convert from a set of vertex data to Self type.

Loading content...