Module amethyst_rendy::sprite

Modules

prefab

Structs

Sprite
Dimensions and texture coordinates of each sprite in a sprite sheet.

SpriteGrid
SpriteGrid controls how a sprite grid is generated when using Sprites::Grid in a SpriteSheetPrefab.

SpriteList
SpriteList controls how a sprite list is generated when using Sprites::List in a SpriteSheetPrefab.

SpritePosition
Represents one sprite in SpriteList. Positions originate in the top-left corner (bitmap image convention).

SpriteRender
Information for rendering a sprite.

SpriteSheet
Meta data for a sprite sheet texture.

SpriteSheetFormat
Allows loading of sprite sheets in RON format.

TextureCoordinates
Texture coordinates of the sprite

Enums

Sprites
Defined the sprites that are part of a SpriteSheetPrefab.

Type Definitions

SpriteSheetHandle
An asset handle to sprite sheet metadata.