Module amethyst_rendy::sprite::prefab

2D Sprite specific prefabs.

Structs

SpriteRenderPrefab
Prefab used to add a sprite to an Entity.

SpriteScenePrefab
Prefab for loading a full scene with sprites.

SpriteSheetLoadedSet
Loaded abstraction for sprite sheets. This is a workaround for prefab humbuggery.

Enums

SpriteSheetPrefab
Defines a spritesheet prefab. Note that this prefab will only load the spritesheet in storage, no components will be added to entities. The add_to_entity will return the Handle<SpriteSheet>. For this reason it is recommended that this prefab is only used as part of other PrefabData or in specialised formats. See SpriteScenePrefab for an example of this.

SpriteSheetReference
Prefab humbuggery. Wraps referencing spritesheets by index or name.