Module amethyst_rendy::submodules

Modules

gather

Structs

DynamicUniform

DynamicVertexData
This structure wraps [PerImageDynamicVertexData], managing multiple instances and providing an easy-to-use interface for having per-image buffers. This is needed because multiple images (frames) can be in flight at any given time, so multiple buffers are needed for the same data.

EnvironmentSub

FlatEnvironmentSub

IndexData
Type used to compile-time specify the type of vertex buffer data managed by a DynamicVertexData

MaterialId

MaterialSub

SkinningSub

TextureId

TextureSub

VertexData
Type used to compile-time specify the type of vertex buffer data managed by a DynamicVertexData

Traits

VertexDataBufferType
Type trait for allowing type-based implementation details for binding the different buffer types of index and vertex DynamicVertexData

Type Definitions

DynamicIndexBuffer
Type alias for a set of dynamic index buffer data to be managed. See the documentation for [DynamicVertexData] for implementation details.

DynamicVertexBuffer
Type alias for a set of dynamic vertex buffer data to be managed. See the documentation for [DynamicVertexData] for implementation details.