Trait amethyst_rendy::system::GraphCreator

pub trait GraphCreator<B: Backend> {
    fn rebuild(&mut self, world: &World) -> bool;

    fn builder(
        &mut self, 
        factory: &mut Factory<B>, 
        world: &World
    ) -> GraphBuilder<B, World>;
}

Graph trait implementation required by consumers. Builds a graph and manages signaling when the graph needs to be rebuilt.

Required methods

fn rebuild(&mut self, world: &World) -> bool

Check if graph needs to be rebuilt. This function is evaluated every frame before running the graph.

fn builder(
&mut self,
factory: &mut Factory<B>,
world: &World
) -> GraphBuilder<B, World>

Retrieve configured complete graph builder.

Loading content...