Trait amethyst_rendy::RenderPlugin

pub trait RenderPlugin<B: Backend>: Debug {
    fn on_plan(
        &mut self, 
        plan: &mut RenderPlan<B>, 
        factory: &mut Factory<B>, 
        world: &World
    ) -> Result<(), Error>;

    fn on_build<'a, 'b>(
        &mut self, 
        _world: &mut World, 
        _builder: &mut DispatcherBuilder<'a, 'b>
    ) -> Result<(), Error> { ... }

    fn should_rebuild(&mut self, _world: &World) -> bool { ... }
}

Basic building block of rendering in [RenderingBundle].

Can be used to register rendering-related systems to the dispatcher, building render graph by registering render targets, adding [RenderableAction]s to them and signalling when the graph has to be rebuild.

Required methods

fn on_plan(
&mut self,
plan: &mut RenderPlan<B>,
factory: &mut Factory<B>,
world: &World
) -> Result<(), Error>

Hook for extending the rendering plan.

Loading content...
Provided methods

fn on_build<'a, 'b>(
&mut self,
_world: &mut World,
_builder: &mut DispatcherBuilder<'a, 'b>
) -> Result<(), Error>

Hook for adding systems and bundles to the dispatcher.

fn should_rebuild(&mut self, _world: &World) -> bool

Hook for providing triggers to rebuild the render graph.

Loading content...