Module amethyst_rendy::visibility

Transparency, visibility sorting and camera centroid culling for 3D Meshes.

Structs

BoundingSphere
Defines a object's bounding sphere used by frustum culling.

Frustum
Simple view Frustum implementation

Visibility
Resource for controlling what entities should be rendered, and whether to draw them ordered or not, which is useful for transparent surfaces.

VisibilitySortingSystem
Determine what entities are visible to the camera, and which are not. Will also sort transparent entities back to front based on distance from camera.